Changed precision to long double

+++ src/mandelb.c
@@ -14,12 +14,12 @@ double scale(int num, int size)
 }
 
 
-int c_get_iterations(double scaled_x, double scaled_y, int max_iters)
+int c_get_iterations(long double scaled_x, long double scaled_y, int max_iters)
 {
     int iters = 0;
 
-    double complex c = scaled_x + scaled_y * I;
-    double complex z = 0.0 + 0.0 * I;
+    long double complex c = scaled_x + scaled_y * I;
+    long double complex z = 0.0 + 0.0 * I;
     
     //try no square root after
     while ((creal(csqrt(z * z)) <= 2) && (iters < max_iters))
